Tried adding some improved DB listener code but commented it out Fixed potential crash involving wrong data class type

+++ sharedCode/src/commonMain/kotlin/sample/networkModels/WeatherResponses.kt
@@ -65,7 +65,7 @@ data class Main(
 
 @Serializable
 data class Wind(
-    val deg: Int,
+    val deg: Double,
     val speed: Double
 )
